TestTaskProfile-3-Set-Up-Authentication - added attributes. Created login endpoint.

+++ TestTaskProfile/Controllers/AccountController.cs
@@ -1,10 +1,12 @@
 ﻿using MediatR;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using TestTaskProfile.CQRS.Users.Commands.CreateUser;
 using TestTaskProfile.CQRS.Users.Commands.DeleteUser;
 using TestTaskProfile.CQRS.Users.Commands.UpdateUser;
 using TestTaskProfile.CQRS.Users.Queries.GetAllUsers;
 using TestTaskProfile.CQRS.Users.Queries.GetUserById;
+using TestTaskProfile.CQRS.Users.Queries.Login;
 using TestTaskProfile.ViewModels.Models;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
@@ -13,11 +15,11 @@ namespace TestTaskProfile.Web.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class UserController : ControllerBase
+    public class AccountController : ControllerBase
     {
         private readonly IMediator _mediator;
 
-        public UserController(IMediator mediator)
+        public AccountController(IMediator mediator)
         {
             _mediator = mediator;
         }
@@ -35,6 +37,14 @@ namespace TestTaskProfile.Web.Controllers
             return await _mediator.Send(new GetUserByIdQuery(id));
         }
 
+        [HttpPost]
+        [Route("login")]
+        public async Task<TokenModel> Login([FromBody] LoginModel model)
+        {
+            var command = new LoginQuery(model);
+            return await _mediator.Send(command);
+        }
+
         // POST api/<CreateUser>
         [HttpPost]
         public async Task<TokenModel> CreateUser([FromBody] CreateUserModel model)
@@ -44,20 +54,48 @@ namespace TestTaskProfile.Web.Controllers
         }
 
         // PUT api/<UserController>/5
+        [Authorize]
         [HttpPut("{id}")]
-        public async Task<GetUserModel> Put(Guid id, [FromBody] UpdateUserModel model)
+        public async Task<IActionResult> Put(Guid id, [FromBody] UpdateUserModel model)
         {
+            var userIdClaim = HttpContext.User.Claims.FirstOrDefault(a => a.Type == "UserId");
+
+            if (string.IsNullOrEmpty(userIdClaim.Value))
+            {
+                return Unauthorized();
+            }
+
+            if(userIdClaim.Value != id.ToString())
+            {
+                return Forbid();
+            }
+
             model.Id = id;
             var updateUserCommandModel = new UpdateUserCommand(model);
-            return await _mediator.Send(updateUserCommandModel);
+            return Ok(await _mediator.Send(updateUserCommandModel));
         }
 
         // DELETE api/<UserController>/5
+        [Authorize]
         [HttpDelete("{id}")]
-        public async Task Delete(Guid id)
+        public async Task<IActionResult> Delete(Guid id)
         {
+            var userIdClaim = HttpContext.User.Claims.FirstOrDefault(a => a.Type == "UserId");
+
+            if (string.IsNullOrEmpty(userIdClaim.Value))
+            {
+                return Unauthorized();
+            }
+
+            if (userIdClaim.Value != id.ToString())
+            {
+                return Forbid();
+            }
+
             var deleteUserCommandModel = new DeleteUserCommand(id);
             await _mediator.Send(deleteUserCommandModel);
+
+            return Ok();
         }
     }
 }
